long is no longer 32 bits, so avoid confusing OF about that. We still have some kind of memory corruption issue that I don't understand when running on my G5.

+++ sys/powerpc/aim/platform_chrp.c
@@ -102,7 +102,7 @@ static u_long
 chrp_timebase_freq(platform_t plat, struct cpuref *cpuref)
 {
 	phandle_t phandle;
-	long ticks = -1;
+	int32_t ticks = -1;
 
 	phandle = cpuref->cr_hwref;
